Move merge and partial response out of shard search
